Focus as focalization_metrics
